--- app/(protected)/complete-profile/page.js
@@ -1,5 +1,0 @@
-import CompleteProfileContent from "@/components/completeProfileContent";
-
-export default async function CompleteProfile() {
-    return <CompleteProfileContent />;
-}
