+++ auth.config.js
@@ -1,10 +1,7 @@
 export const authConfig = {
     publicRoutes: ["/", "/auth/confirm", "/error"],
-    protectedRoutes: ["/details", "/rsvp", "/rsvp/thanks", "/complete-profile"],
+    protectedRoutes: ["/details", "/rsvp", "/rsvp/thanks"],
     unAuthedHomeRoute: "/",
     authedHomeRoute: "/details",
     ticketedRoute: ["/rsvp/thanks"],
-    emailOnlyRoutes: ["/rsvp"],
-    noEmailUpdateRoute: "/complete-profile",
-    noEmailPlaceHolder: 'guest.ramyandshazia.com'
 };
