+++ components/DialogEmailInput.jsx
@@ -13,12 +13,9 @@ import {
     DialogTitle,
     DialogTrigger,
 } from "@/components/ui/dialog";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { toast } from "sonner";
-import {
-    sendMagicLinkEmail,
-    sendMagicLinkTextNoEmail,
-} from "@/app/auth/confirm/_lib/actions";
+import { sendMagicLink } from "@/app/auth/confirm/_lib/actions";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
@@ -28,75 +25,23 @@ import {
     FormItem,
     FormControl,
     FormLabel,
-    FormDescription,
 } from "@/components/ui/form";
-import parsePhoneNumber from "libphonenumber-js";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
-const emailSchema = z.object({
+const schema = z.object({
     email: z.email("Invalid email address"),
 });
 
-const textSchema = z.object({
-    phone: z.string().refine(
-        (val) => {
-            const phone = parsePhoneNumber(val, "US");
-            if (phone?.isValid()) {
-                return true;
-            }
-            return false;
-        },
-        {
-            message: "Please enter a valid phone number",
-        }
-    ),
-});
-
 const DialogEmailInput = () => {
-    const emailForm = useForm({
-        resolver: zodResolver(emailSchema),
+    const form = useForm({
+        resolver: zodResolver(schema),
         defaultValues: { email: "" },
     });
 
-    const textForm = useForm({
-        resolver: zodResolver(textSchema),
-        defaultValues: { phone: "" },
-    });
-
     const [open, setOpen] = useState(false);
-    const [isCoolingDown, setIsCoolingDown] = useState(false);
-    const [countdown, setCountdown] = useState(120);
 
-    useEffect(() => {
-        if (!isCoolingDown) return;
-
-        if (countdown <= 0) {
-            setIsCoolingDown(false);
-            setCountdown(120); // Reset for next time
-            return;
-        }
-
-        const timerId = setInterval(() => {
-            setCountdown((prevCountdown) => prevCountdown - 1);
-        }, 1000);
-
-        // Cleanup interval on component unmount or state change
-        return () => clearInterval(timerId);
-    }, [isCoolingDown, countdown]);
-
-    const formatTime = (seconds) => {
-        const minutes = Math.floor(seconds / 60);
-        const remainingSeconds = seconds % 60;
-        return `${minutes}:${remainingSeconds.toString().padStart(2, "0")}`;
-    };
-
-    const startCooldown = () => {
-        setIsCoolingDown(true);
-    };
-
-    const onSubmitEmail = async ({ email }) => {
+    const onSubmit = async ({ email }) => {
         const sanitizedEmail = email.trim().toLowerCase();
-        const emailSuccess = await sendMagicLinkEmail(sanitizedEmail);
+        const emailSuccess = await sendMagicLink(sanitizedEmail);
         if (emailSuccess?.error) {
             toast.error("Error sending link", {
                 description: emailSuccess.error,
@@ -107,40 +52,14 @@ const DialogEmailInput = () => {
                 description: "Check your email!",
                 position: "bottom-right",
             });
-            startCooldown();
         }
-    };
 
-    const handleOpenChange = (isOpen) => {
-        setOpen(isOpen);
-        if (!isOpen) {
-            setIsCoolingDown(false);
-            setCountdown(120);
-            emailForm.reset();
-            textForm.reset();
-        }
-    };
-
-    const onSubmitText = async ({ phone }) => {
-        const sanitizedPhone = phone.trim().toLowerCase();
-        const textSuccess = await sendMagicLinkTextNoEmail(sanitizedPhone);
-        if (textSuccess?.error) {
-            toast.error("Error sending link", {
-                description: textSuccess.error,
-                position: "bottom-right",
-            });
-        } else {
-            toast.success("Link Sent", {
-                description: "Check your texts!",
-                position: "bottom-right",
-            });
-
-            startCooldown();
-        }
+        form.reset();
+        setOpen(false);
     };
 
     return (
-        <Dialog open={open} onOpenChange={handleOpenChange}>
+        <Dialog open={open} onOpenChange={setOpen}>
             <DialogTrigger asChild>
                 <Button
                     variant={"outline"}
@@ -152,173 +71,54 @@ const DialogEmailInput = () => {
                 </Button>
             </DialogTrigger>
             <DialogContent>
-                <Tabs defaultValue="email">
-                    <div
-                        className={cn(
-                            "container flex flex-row justify-center items-center mb-5"
-                        )}
-                    >
-                        <TabsList className={cn("")}>
-                            <TabsTrigger value="email">Email</TabsTrigger>
-                            <TabsTrigger value="text">Phone</TabsTrigger>
-                        </TabsList>
-                    </div>
-                    <TabsContent value="email">
-                        <Form {...emailForm}>
-                            <form
-                                onSubmit={emailForm.handleSubmit(onSubmitEmail)}
+                <Form {...form}>
+                    <form onSubmit={form.handleSubmit(onSubmit)}>
+                        <DialogHeader>
+                            <DialogTitle> Get Access and RSVP</DialogTitle>
+                            <DialogDescription>
+                                We are limiting access to invited guests only
+                                please enter your email and a link to enter will
+                                be emailed to you
+                            </DialogDescription>
+                        </DialogHeader>
+                        <div className="grid gap-4 mt-5">
+                            <div className="grid gap-3">
+                                <FormField
+                                    control={form.control}
+                                    name="email"
+                                    render={({ field }) => (
+                                        <FormItem>
+                                            <FormLabel> Email </FormLabel>
+                                            <FormControl>
+                                                <Input
+                                                    type="email"
+                                                    placeholder="you@example.com"
+                                                    {...field}
+                                                />
+                                            </FormControl>
+                                        </FormItem>
+                                    )}
+                                />
+                            </div>
+                        </div>
+                        <DialogFooter className={cn("mt-5")}>
+                            <DialogClose asChild>
+                                <Button type="button" variant="outline">
+                                    Close
+                                </Button>
+                            </DialogClose>
+                            <Button
+                                type="submit"
+                                variant="secondary"
+                                disabled={form.formState.isSubmitting}
                             >
-                                <DialogHeader>
-                                    <DialogTitle>
-                                        {" "}
-                                        Get Access and RSVP
-                                    </DialogTitle>
-                                    <DialogDescription
-                                        className={cn("text-justify")}
-                                    >
-                                        We are limiting access to invited guests
-                                        only. Please enter your email address
-                                        and a link will be sent to you
-                                    </DialogDescription>
-                                </DialogHeader>
-                                <div className="grid gap-4 mt-5">
-                                    <div className="grid gap-3">
-                                        <FormField
-                                            control={emailForm.control}
-                                            name="email"
-                                            render={({ field, fieldState }) => (
-                                                <FormItem>
-                                                    <FormLabel>
-                                                        {" "}
-                                                        Email{" "}
-                                                    </FormLabel>
-                                                    <FormControl>
-                                                        <Input
-                                                            type="email"
-                                                            placeholder="you@example.com"
-                                                            {...field}
-                                                        />
-                                                    </FormControl>
-                                                    {fieldState.error && (
-                                                        <FormDescription
-                                                            className={cn(
-                                                                "text-red-500"
-                                                            )}
-                                                        >
-                                                            {
-                                                                fieldState.error
-                                                                    .message
-                                                            }
-                                                        </FormDescription>
-                                                    )}
-                                                </FormItem>
-                                            )}
-                                        />
-                                    </div>
-                                </div>
-                                <DialogFooter className={cn("mt-5")}>
-                                    <DialogClose asChild>
-                                        <Button type="button" variant="outline">
-                                            Close
-                                        </Button>
-                                    </DialogClose>
-                                    <Button
-                                        type="submit"
-                                        variant="secondary"
-                                        disabled={
-                                            emailForm.formState.isSubmitting ||
-                                            isCoolingDown
-                                        }
-                                    >
-                                        {isCoolingDown
-                                            ? `Resend in ${formatTime(countdown)}`
-                                            : emailForm.formState.isSubmitting
-                                              ? "Sending..."
-                                              : "Send Link"}
-                                    </Button>
-                                </DialogFooter>
-                            </form>
-                        </Form>
-                    </TabsContent>
-                    <TabsContent value="text">
-                        <Form {...textForm}>
-                            <form
-                                onSubmit={textForm.handleSubmit(onSubmitText)}
-                            >
-                                <DialogHeader>
-                                    <DialogTitle>
-                                        {" "}
-                                        Get Access and RSVP
-                                    </DialogTitle>
-                                    <DialogDescription
-                                        className={cn("text-justify")}
-                                    >
-                                        We are limiting access to invited guests
-                                        only. Please enter your phone number and
-                                        a link to enter will be sent to you via
-                                        text.
-                                    </DialogDescription>
-                                </DialogHeader>
-                                <div className="grid gap-4 mt-5">
-                                    <div className="grid gap-3">
-                                        <FormField
-                                            control={textForm.control}
-                                            name="phone"
-                                            render={({ field, fieldState }) => (
-                                                <FormItem>
-                                                    <FormLabel>
-                                                        {" "}
-                                                        Phone{" "}
-                                                    </FormLabel>
-                                                    <FormControl>
-                                                        <Input
-                                                            type="phone"
-                                                            placeholder="123-456-7890"
-                                                            {...field}
-                                                        />
-                                                    </FormControl>
-                                                    {fieldState.error && (
-                                                        <FormDescription
-                                                            className={cn(
-                                                                "text-red-500"
-                                                            )}
-                                                        >
-                                                            {
-                                                                fieldState.error
-                                                                    .message
-                                                            }
-                                                        </FormDescription>
-                                                    )}
-                                                </FormItem>
-                                            )}
-                                        />
-                                    </div>
-                                </div>
-                                <DialogFooter className={cn("mt-5")}>
-                                    <DialogClose asChild>
-                                        <Button type="button" variant="outline">
-                                            Close
-                                        </Button>
-                                    </DialogClose>
-                                    <Button
-                                        type="submit"
-                                        variant="secondary"
-                                        disabled={
-                                            textForm.formState.isSubmitting ||
-                                            isCoolingDown
-                                        }
-                                    >
-                                        {isCoolingDown
-                                            ? `Resend in ${formatTime(countdown)}`
-                                            : textForm.formState.isSubmitting
-                                              ? "Sending..."
-                                              : "Send Link"}
-                                    </Button>
-                                </DialogFooter>
-                            </form>
-                        </Form>
-                    </TabsContent>
-                </Tabs>
+                                {form.formState.isSubmitting
+                                    ? "Sending..."
+                                    : "Send Magic Link"}
+                            </Button>
+                        </DialogFooter>
+                    </form>
+                </Form>
             </DialogContent>
         </Dialog>
     );
