+++ components/optInForm.jsx
@@ -71,6 +71,7 @@ const OptInForm = () => {
 
         form.reset();
         toast.info("Your submission has successfully been recorded!");
+
         redirect("/");
     };
