+++ package.json
@@ -21,7 +21,6 @@
     "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-switch": "^1.2.5",
-    "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.7",
     "@react-email/components": "0.3.3",
     "@supabase/ssr": "^0.6.1",
@@ -30,7 +29,6 @@
     "clsx": "^2.1.1",
     "dotenv": "^17.2.1",
     "framer-motion": "^12.15.0",
-    "libphonenumber-js": "^1.12.12",
     "lucide-react": "^0.511.0",
     "motion": "^12.19.1",
     "next": "15.3.2",
@@ -44,7 +42,6 @@
     "sonner": "^2.0.6",
     "tailwind-merge": "^3.3.0",
     "three": "^0.177.0",
-    "twilio": "^5.8.0",
     "uuid": "^11.1.0",
     "vanta": "^0.5.24",
     "zod": "^4.0.10"
